backend/requirements.txt
anyascii==0.3.1
APScheduler==3.9.1
asgiref==3.5.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
dj-static==0.0.6
Django==4.0.4
django-allauth==0.50.0
django-apscheduler==0.6.2
django-cors-headers==3.12.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.3
django-ninja==0.17.0
django-rest-framework==0.1.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
ipython==8.3.0
jedi==0.18.1
l18n==2021.3
matplotlib-inline==0.1.3
oauthlib==3.2.0
openpyxl==3.0.9
orjson==3.6.8
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
static3==0.7.0
tablib==3.2.1
telepath==0.2
toml==0.10.2
traitlets==5.2.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wagtail==2.16.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.1.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
